+++ package.json
@@ -20,6 +20,7 @@
     "exif": "^0.6.0",
     "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
+    "move-file": "^1.0.0",
     "mv": "^2.1.1",
     "simple-node-logger": "^0.93.37",
     "strip-ansi": "^4.0.0"
